Add separate tests for violations controlled by 'i_control_code' option
