+++ tests/test_checker/test_noqa.py
@@ -228,6 +228,13 @@ SHOULD_BE_RAISED = types.MappingProxyType({
     'WPS702': 1,
 })
 
+SHOULD_BE_RAISED_NO_CONTROL = types.MappingProxyType({
+    'WPS113': 1,
+
+    'WPS412': 0,
+    'WPS413': 0,
+})
+
 
 def _assert_errors_count_in_output(output, errors, all_violations):
     found_errors = Counter(
@@ -269,6 +276,36 @@ def test_noqa_fixture_disabled(absolute_path, all_violations):
 
     _assert_errors_count_in_output(stdout, SHOULD_BE_RAISED, all_violations)
     assert len(SHOULD_BE_RAISED) == len(all_violations)
+
+
+def test_noqa_fixture_disabled_no_control(
+    absolute_path,
+    all_controlled_violations,
+):
+    """End-to-End test to check rules controlled by `i_control_code` option."""
+    process = subprocess.Popen(
+        [
+            'flake8',
+            '--i-dont-control-code',
+            '--disable-noqa',
+            '--isolated',
+            '--select',
+            'WPS',
+            absolute_path('fixtures', 'noqa_controlled.py'),
+        ],
+        stdout=subprocess.PIPE,
+        stderr=subprocess.PIPE,
+        universal_newlines=True,
+        encoding='utf8',
+    )
+    stdout, _ = process.communicate()
+
+    _assert_errors_count_in_output(
+        stdout,
+        SHOULD_BE_RAISED_NO_CONTROL,
+        all_controlled_violations,
+    )
+    assert len(SHOULD_BE_RAISED_NO_CONTROL) == len(all_controlled_violations)
 
 
 def test_noqa_fixture(absolute_path):
